Make ScopedURLFetcherFactory sequenced-safe.
This CL allows ScopedURLFetcherFactory to be used in a test
environment exposing a sequenced task runner instead of
requiring a single thread task runner, a relaxation of the
requirements (single thread is a stricter instance of a
sequened task runner).

The URL fetcher code has already been relaxed in a previous
CL to only require a sequenced runner, but the corresponding
update to ScopedURLFetcherFactory was overlooked.

BUG=675631

Review-Url: https://codereview.chromium.org/2757233003
Cr-Original-Commit-Position: refs/heads/master@{#458201}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 770b36c2734d41edbb8d5b260bcfc9c29fe16afd