Reporting: Implement header parser.
Reporting is a spec for delivering out-of-band reports from various
other parts of the browser. See http://wicg.github.io/reporting/ for
the spec, or https://goo.gl/pygX5I for details of the planned
implementation in Chromium.

This implements the header parser, which parses Report-To headers
and applies the correct updates to the Reporting cache.

BUG=704259

Review-Url: https://codereview.chromium.org/2689953004
Cr-Original-Commit-Position: refs/heads/master@{#460434}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1690bc6eba0b829b5250c2e26347a951ed653515