Reporting: Implement uploader.
Reporting is a spec for delivering out-of-band reports from various
other parts of the browser. See http://wicg.github.io/reporting/.

The uploader handles the network side of report delivery: it takes the
assembled JSON and an endpoint URL, and handles sending the URL request
and categorizing the response according to the Reporting spec.

BUG=704259

Review-Url: https://codereview.chromium.org/2723563002
Cr-Original-Commit-Position: refs/heads/master@{#460846}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2f46b5780970e186eda8641be9138ec551e3f1db