Remove URLRequestJobFactory::IsHandledURL.
In practice, this was just a combination of GURL::is_valid() and
URLRequestJobFactory::IsHandledProtocol(GURL::is_scheme()).

This is just motivated by a desire to simplify the knobs exposed by
the network stack, for servicification.  This does add a number of
GURL::is_valid() checks, but I suspect most of them will go away, as we
adapt the launch externally handled protocol path.

BUG=706942

Review-Url: https://codereview.chromium.org/2783343002
Cr-Original-Commit-Position: refs/heads/master@{#461518}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1918ae78a8154c4dfc31fd238f9a00c53bbd96fd