Reporting: Wrap existing classes in context.
Reporting is a spec for delivering out-of-band reports from various
other parts of the browser. See http://wicg.github.io/reporting/ for
the spec, or https://goo.gl/pygX5I for details of the planned
implementation in Chromium.

This CL takes the pile of small classes that I've written so far and
wraps them in a single "context" object.

It also changes the classes to accept a pointer to the context instead
of several pointers to individual dependencies.

It *also* creates a ReportingTestBase class that provides a properly
configured context for all of the existing unittests to use, and
switches the tests over.

BUG=704259

Review-Url: https://codereview.chromium.org/2751103002
Cr-Original-Commit-Position: refs/heads/master@{#462901}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ee4b55e3526065ac1d0f09a80372e17f26892c8e