Reporting: Wrap context in service.
Reporting is a spec for delivering out-of-band reports from various
other parts of the browser. See http://wicg.github.io/reporting/ for
the spec, or https://goo.gl/pygX5I for details of the planned
implementation in Chromium.

This wraps the context in a service that keeps the implementation
details of Reporting private while implementing and exposing the
high-level operations that other components require from it.

BUG=704259

Review-Url: https://codereview.chromium.org/2770443002
Cr-Original-Commit-Position: refs/heads/master@{#462943}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 381d77efd050a4ef04eba2ce52f06741ac0d1107