Link NetLogs for HttpStreamFactoryImpl::Job and existing QuicStreamFactory::Job
When HttpStreamFactoryImpl::Job init connection to QUIC server, in some
cases, it might not result in QuicStreamFactory creating a new
QuicStreamFactory::Job to serve the QuicRequest but use an exisitng
QuicStreamFactory::Job. Link the HttpStreamFactoryImpl::Job and
QuicStreamFactory::Job so that the HttpStreamFactoryImpl::Job::DoInitConnection
is traceable.

BUG=702411

Review-Url: https://codereview.chromium.org/2811573003
Cr-Original-Commit-Position: refs/heads/master@{#463536}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 011ef8d97d0b06877ca4cfc621d2e56f19aeef8c