Reporting: Observe network changes and clear reports/clients
Reporting is a spec for delivering out-of-band reports from various
other parts of the browser. See http://wicg.github.io/reporting/ for
the spec, or https://goo.gl/pygX5I for details of the planned
implementation in Chromium.

This CL adds a ReportingNetworkChangeObserver that can clear reports
and/or clients (as configured in ReportingPolicy) on network changes.

BUG=704259

Review-Url: https://codereview.chromium.org/2813103005
Cr-Original-Commit-Position: refs/heads/master@{#466113}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 164e4cdfd53297fe3801ae893bbaa48651e1b339