Fix incorrect computation of LastModified in SimpleCache, avoid making useless syscalls[1]. Those particular ones seem to typically be ultra cheap, but why do them when not needed?
The code was written to use simple_util::GetMTime because at the time it
was more precise than File::GetInfo when it came to timestamps; trouble
is, path_ is the cache directory, not the cache entry file (see e.g.
GetFilenameFromFileIndex), so this is the last modified timestamp
for the wrong thing; and as far as I can see File::GetInfo is as precise
these days.

As far as I can see w/grep and codesearch, nothing ultimately uses
DiskCache::Entry::GetLastModified() outside tests, so the biggest
impact was likely that SyncOpenEntryAge stat collected was bogus.

[1] Those were the looks up of info on the cache directory we saw when
 looking at syscall trace on Windows.

BUG=714143

Review-Url: https://codereview.chromium.org/2791493003
Cr-Original-Commit-Position: refs/heads/master@{#466674}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2cac5e68544bf2f5fbaec00f62ffe0c6e7957598