Sync the browser test host resolver with the network process before a test runs.
This fixes the browser tests which use the in-process host resolver, which depends on globals in the network code.

Most of the tests already call host_resolver()->AddRule in their SetUpOnMainThread. This will be enforced once the remaining ones which don't do this are updated.

BUG=713847

Review-Url: https://codereview.chromium.org/2837813004
Cr-Original-Commit-Position: refs/heads/master@{#467082}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 700d86f404cceb4d49e2a1ed0b6f584839e43fde