Add UMA to track the number of live URLRequests.
UMA data shows that there can be a large number of active
HttpStreamRequests at a given time. The number reported by
ResourceDispatcherHostImpl doesn't account for the large number of
HttpStreamRequests. This CL adds a histogram to track the number of
active URLRequests, so we can know whether the cause for the large
number of HttpStreamRequests is internal or external to the network
stack.

BUG=711721

Review-Url: https://codereview.chromium.org/2837313002
Cr-Original-Commit-Position: refs/heads/master@{#467332}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3fcbcc190e8d78af50840b7535893a8662b5e74b