Remove URLRequestJob::GetResponseCode implementations from content, chrome, and net/.
URLRequestJob itself now has a default implementation, which gets
the response code from the headers returned by GetResponseInfo.
We eventually want to remove this method, since it's redundant,
and has weird behavior for just when it is and is not safe to
call.

BUG=688481

Review-Url: https://codereview.chromium.org/2835533002
Cr-Original-Commit-Position: refs/heads/master@{#467839}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e47392b3d5fa37b785ea7494bc7bd32163ba9188