Remove client_certs from SSLCertRequestInfo.
The client_certs aren't actually part of the certificate request, rather the client_certs member was used to store the list of matching certificates on the client side and pass them through to the certificate selector.
Pass the list of certs through the relevant callbacks instead.

BUG=166642,394131

Review-Url: https://codereview.chromium.org/2838243002
Cr-Original-Commit-Position: refs/heads/master@{#467901}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7ed243fc89fb72230c47f59742128add62ae1d83