Reporting: Set up in ProfileImplIOData.
Reporting is a spec for delivering out-of-band reports from various
other parts of the browser. See http://wicg.github.io/reporting/ for
the spec, or https://goo.gl/pygX5I for details of the planned
implementation in Chromium.

This CL creates a barebones ChromeReportingDelegate, initializes the
ReportingService in ProfileImplIOData::InitializeInternal, and sets the
service on the URLRequestContext.

BUG=704259

Review-Url: https://codereview.chromium.org/2769083005
Cr-Original-Commit-Position: refs/heads/master@{#468038}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1d92f015d0b85e071f0040c53ccd8f145b55e4c1