Allow ProxyService to share URLRequestContext with everything else.
This CL makes ProxyScriptFetcherImpl use highest priority and
LOAD_IGNORE_LIMITS, and adds an OnShutdown method to ProxyService,
which must be called before teardown when the ProxyService makes
requests using a URLRequestContext that it is also used with.

No consumer currently uses the new functionality.  I plan to
switch IOThread/ProfileIOData to using it in another CL, and
add a call in URLRequestContextStorage's destructor.  May also
be worth making it mandatory as well, though that will involve
more changes.

BUG=715697

Review-Url: https://codereview.chromium.org/2845643003
Cr-Original-Commit-Position: refs/heads/master@{#469005}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ed8d7e4383d5dbf7e94291b7960b5a4616b4a45e