Use SSL_set_strict_cipher_list and prune removed things.
Also check for errors in setting the cipher list. Not all failure cases
are the empty cipher list. Since that can only happen if the user passed
a bunch of (unsupported) command-line exclusion flags, it seems not
worth the trouble to detect that case and let it fall-through.

BUG=none

Review-Url: https://codereview.chromium.org/2855883006
Cr-Original-Commit-Position: refs/heads/master@{#469097}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1863716bb870b7b229530f2bdf853322b30178c0