The Handle is owned by the holder of the Handle and can outlive the session. In the event that it does outlive the session, it will record state from the session in member variables so that the holder of a session need not be aware of the lifetime of the underlying session.
As a consequence, this change eliminates the need for the QuicChromiumClientSession::Observer interface, and hence eliminates the upcalls from the session to the streams. There are still upcalls from the QuicStream to the QuicHttpStream (and friends) but those can be removed in subsequent CLs by following the same pattern.

BUG=716563

Review-Url: https://codereview.chromium.org/2334943002
Cr-Original-Commit-Position: refs/heads/master@{#469756}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f0b18c8a8efc9b2dd335a0b29bd243b0711919a6