Fix QuicNetworkTransactionTest.RetryMisdirectedRequest
This test was depending on the SSL deprecated cipher fallback (which is
a quirk to probe server cipher preferences and not something QUIC may
depend on) to clear away the ERR_CONNECTION_CLOSED when it shouldn't
have been surfacing one to begin with.

Fortunately, the actual logic isn't relying on this. The test just needs
fixing. Rather than scheduling three connections, schedule two. The main
job hangs until after all the QUIC machinery is done, then we unblock
the main job and let that complete.

BUG=546991,684730

Review-Url: https://codereview.chromium.org/2856073003
Cr-Original-Commit-Position: refs/heads/master@{#469799}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a4449725c3c1d07ecd8008d5d03227c4588b85dc