Buffer initial response headers in QuicChromiumClientStream if headers arrive on a pushed stream before the delegate is set.
Eliminate the "delegate tasks" member. It turns out that there can
be at most one delegate task, because OnDataAvailable is only
invoked on the delegate after the headers are delivered to it. So
only OnHeadersAvailable was ever buffered. This CL just makes that
explicit.

Review-Url: https://codereview.chromium.org/2865893002
Cr-Original-Commit-Position: refs/heads/master@{#470226}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6788ad5e2f48d4228195884715f1809e2922e3ac