Reporting: Document browsing data remover + endpoint backoff
The ReportingBrowsingDataRemover currently doesn't remove entries from
the endpoint backoff map in the ReportingEndpointManager, because it's
never persisted to disk. This CL adds comments in a couple places to
note this, so someone implementing persistence for that knows to
implement browsing data removal as well.

BUG=704259,676324

Review-Url: https://codereview.chromium.org/2862003003
Cr-Original-Commit-Position: refs/heads/master@{#470336}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cf0772c92c9909b3ad0f107541a533c567bfda75