Unit-test CAPI and CNG SSLPrivateKey implementations.
This does not cover the CryptAcquireCertificatePrivateKey call (from my
manual testing, it didn't appear that the imported keys showed up
via CryptAcquireCertificatePrivateKey at all), but it tests that we are calling
CAPI and CNG correctly for the algorithms we are trying to implement. This
will be useful when we revise the interface for RSA-PSS and add support for
it. Also, it means we actually have some test for this stuff!

BUG=673058

Review-Url: https://codereview.chromium.org/2865883002
Cr-Original-Commit-Position: refs/heads/master@{#470612}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d6d1e4d17563f475169ea8564be5e581759eb8fd