Add const to SpdyFrameIR in SpdyFrameSequence.
This is to reflect that after the creation of a SpdyFrameSequence derived
class, the SpdyFrameIR should not be mutated.

This CL lands server change 152709726 by yasong.

Also fix formatting and comments in a number of places for consistency
with internal code.

Also use SpdyMakeUnique and SpdyWrapUnique instead of their base::
counterparts.

BUG=488484

Review-Url: https://codereview.chromium.org/2876613002
Cr-Original-Commit-Position: refs/heads/master@{#470963}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 07424d1b2e8ac5a7c1a9ed3079e21e52d9a13085