Fix SpdySessionKey for HTTP/2 alternative Jobs.
Change HttpStreamFactoryImpl::Job::GetSpdySessionKey() to construct a
SpdySessionKey from the request URL, not the destination.  This is the
correct behavior, because the only layers that should worry about the
destination are the socket pool and below.

This does not change behavior in production, because
HostPortPair::FromURL(origin_url_) can only differ from |destination_|
for alternative jobs, and HTTP/2 alternative jobs are disabled in
https://crrev.com/2821463002.

BUG=615413, 706974

Review-Url: https://codereview.chromium.org/2887773006
Cr-Original-Commit-Position: refs/heads/master@{#472919}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1fc76912bc055188dc76c09f6032096c079d4d86