Fix comment about "happy eyeballs" IPv4 fallback
The comment used to claim that in an address list with both IPv6 and IPv4
addresses, the IPv6 addresses will come first, which is usually but not
always true. Now it correctly explains that the fallback is only used
when an IPv6 address is first.

Review-Url: https://codereview.chromium.org/2895633002
Cr-Original-Commit-Position: refs/heads/master@{#473351}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2a552a75bf20b63dc035145a8b2eaef54c1e034c