Allow use of Mojo/V8 ProxyResolvers with URLRequestContextBuilder.
This adds URLRequestContextBuilderV8, which can construct a
URLRequestContext using either type of ProxyResolver. When using such
a resolver, the ProxyService will fetch requests using the
URLRequestContext that was created with the builder.

Added a subclass instead of modifying URLRequestContextBuilder so it
can continue to be used without depending on V8.

This was originally landed in
https://codereview.chromium.org/2881613002/, and reverted in
https://codereview.chromium.org/2888043008/.  This version of the CL
fixes a test-only lifetime issue.

BUG=717253, 724471

Review-Url: https://codereview.chromium.org/2898073002
Cr-Original-Commit-Position: refs/heads/master@{#473933}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 93be9ca690ac29ad965a7f7bf6f899cf05c0c582