Un-revert SQLiteChannelIDStore threshold changes
The change at crrev.com/473631 reverted a change to the thresholds in
SQLiteChannelIDStore because that change was supposed to guarantee that
Channel IDs get flushed when Cookies get flushed. However, that CL was
incomplete.

This change in the thresholds is temporary until the work in that
previous CL can be fixed.

Bug: 548423
Change-Id: I103d67eb5aa34a209f08656d724ef33efe82898c
Reviewed-on: https://chromium-review.googlesource.com/514147
Reviewed-by: Matt Mueller <mattm@chromium.org>
Commit-Queue: Nick Harper <nharper@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#474472}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2e025bde51430446f859a81d274e9ae721e47555