Partial Protobuf for Traffic Annotation.
In many use cases, it's better to define some part of a network traffic
annotation in one function and the reset of it in another one. like when a
function makes the initial request and controls policies, and then passes it
to another one to set/unset the cookies and perform the actual request. This CL
adds required definitions to make partial annotations.

BUG=656607

patch from issue 2815293002 at patchset 160001 (http://crrev.com/2815293002#ps160001)

Review-Url: https://codereview.chromium.org/2869373002
Cr-Original-Commit-Position: refs/heads/master@{#474602}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 99c6fcef9cff05b0df3a536623f8d8251c2f2a6b