Do not use ScopedTaskEnvironment::RunUntilIdle() in URLRequestSimpleJobTest.
URLRequestSimpleJobTest.CancelAfterFirstReadStarted doesn't need
ScopedTaskEnvironment::RunUntilIdle() because it doesn't depend on
tasks running in TaskScheduler.

Moving from ScopedTaskEnvironment::RunUntilIdle() to
RunLoop::RunUntilIdle() in this test is important because the
test posts a MessageLoop::QuitWhenIdleClosure() (via TestDelegate),
which is incompatible with changes that we want to make to
ScopedTaskEnvironment::RunUntilIdle().

BUG=724077

Review-Url: https://codereview.chromium.org/2903213003
Cr-Original-Commit-Position: refs/heads/master@{#474772}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 38ca4d86dba6022a465b9966e6917a9e004e919b